added view for top-rated movies and link to discover page
