+++ app/controllers/movies_controller.rb
@@ -3,6 +3,7 @@ class MoviesController < ApplicationController
    @user = User.find(params[:id])
   end
   def index
-    
+    @movies = MoviesSearch.new.movies
+    @user = User.find(params[:id])
   end
 end
