+++ app/models/viewing_party.rb
@@ -21,7 +21,7 @@ class ViewingParty < ApplicationRecord
 
   def image 
     #this will return the url path for the movie image for that party
-    movie = MovieSearch.new.movie(movie_id)
+    movie = MovieDetailsSearch.new.movie(movie_id)
     config_path = ConfigSearch.new.base_path
     "#{config_path}w200#{movie.partial_image_path}"
   end
