+++ app/poros/config_search.rb
@@ -1,4 +1,4 @@
-require "./app/poros/movie"
+
 require "./app/services/tmdb_service"
 
 class ConfigSearch
