+++ app/poros/movie.rb
@@ -1,12 +1,10 @@
 class Movie 
-  attr_reader :partial_image_path, :title, :runtime, :genre, :vote, :summary
+  attr_reader :id, :title, :vote
 
   def initialize(data)
-    @partial_image_path = data[:poster_path]
+    @id = data[:id]
     @title = data[:original_title]
-    @runtime = data[:runtime]
-    @genres = data[:genres].map { |genre| genre[:name] }
     @vote = data[:vote_average]
-    @summary = data[:overview]
+    
   end
 end
