+++ app/views/movies/discover.html.erb
